moved Scheduler and SchedulerModule to connector, added intentions
