+++ api/src/main/scala/quasar/api/scheduler/SchedulerError.scala
@@ -21,10 +21,9 @@ import slamdata.Predef._
 import cats.Show
 import cats.implicits._
 
-sealed trait SchedulingError[+I, +II, +C, +CC] extends Product with Serializable
+sealed trait SchedulerError[+I, +C] extends Product with Serializable
 
 object SchedulerError {
-  sealed trait SchedulerError[+I, +C] extends SchedulingError[I, Nothing, C, Nothing]
   sealed trait CreateError[+C] extends SchedulerError[Nothing, C]
 
   final case class SchedulerUnsupported(kind: SchedulerType, supported: Set[SchedulerType])
@@ -47,10 +46,6 @@ object SchedulerError {
       extends SchedulerError[I, Nothing]
   final case class DeleteError[C](config: C)
       extends SchedulerError[Nothing, C]
-
-  sealed trait IntentionError[+I, +C] extends SchedulingError[Nothing, I, Nothing, C]
-  final case class IncorrectIntention[C](config: C) extends IntentionError[Nothing, C]
-  final case class IntentionNotFound[I](index: I) extends IntentionError[I, Nothing]
 
   object CreateError {
     implicit def show[C: Show]: Show[CreateError[C]] = Show.show {
@@ -80,17 +75,4 @@ object SchedulerError {
       case DeleteError(config) => s"DeleteError(${config.show})"
     }
   }
-
-  object IntentionError {
-    implicit def show[I: Show, C: Show]: Show[IntentionError[I, C]] = Show.show {
-      case IncorrectIntention(config) => s"IncorrectIntention(${config.show})"
-      case IntentionNotFound(index) => s"IntentionNotFound(${index.show})"
-    }
-  }
-
-  implicit def show[I: Show, II: Show, C: Show, CC: Show]: Show[SchedulingError[I, II, C, CC]] =
-    Show.show {
-      case e: SchedulerError[I, C] => e.show
-      case e: IntentionError[II, CC] => e.show
-    }
 }
