+++ impl/src/main/scala/quasar/impl/scheduler/SchedulerModules.scala
@@ -21,6 +21,7 @@ import slamdata.Predef._
 import quasar.api.scheduler._
 import quasar.api.scheduler.SchedulerError._
 import quasar.connector.MonadResourceErr
+import quasar.connector.scheduler.{Scheduler, SchedulerModule}
 import quasar.impl.IncompatibleModuleException.linkScheduler
 
 import argonaut.Json
